+++ inst/runit_tests/runit-get_boundary_radius.R
@@ -1,8 +1,24 @@
 if (interactive()) pkgload::load_all(".")
 test_get_boundary_radius <- function() {
     expectation <- 12.625
-    result <-  fritools::strip_off_attributes(get_boundary_radius(505,
-                                                                  unit = "m"))
+    dbh <- 505
+    boundary_radius <- get_boundary_radius(dbh, unit = "m", is_ti_round = FALSE) 
+    result <- fritools::strip_off_attributes(boundary_radius)
+    RUnit::checkIdentical(result, expectation)
+
+    # no rounding:
+    expectation <- 1262.5
+    dbh <- 505
+    boundary_radius <- get_boundary_radius(dbh, unit = "cm",
+                                           is_ti_round = FALSE) 
+    result <- fritools::strip_off_attributes(boundary_radius)
+    RUnit::checkIdentical(result, expectation)
+
+    # rounding to even:
+    expectation <- 1262
+    dbh <- 505
+    boundary_radius <- get_boundary_radius(dbh, unit = "cm", is_ti_round = TRUE) 
+    result <- fritools::strip_off_attributes(boundary_radius)
     RUnit::checkIdentical(result, expectation)
 }
 if (interactive()) {