+++ inst/tinytest/test_check_boundaries.R
@@ -4,6 +4,6 @@ if (interactive()) {
 }
 
 expectation <- 12.625
-br <- treePlotArea::get_boundary_radius(505, unit = "m")
+br <- treePlotArea::get_boundary_radius(505, unit = "m", is_ti_round = FALSE)
 result <- as.numeric(br)
 expect_identical(result, expectation)